# contracts/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Policy Orchestrator & Algorithm Execution API
  version: 1.0.0
  description: >-
    Enterprise-grade, vendor-agnostic REST API for repository invariant auditing,
    grounded hybrid GraphRAG, autonomous AI Agent orchestration, and direct execution
    of all 152 Layer 1 algorithms across Search, Observability, Update, Vector, Graph, Filter, and Transform categories.
    Strictly conforms to the RFC API Request & Response Structure Specification (v5.0).
servers:
  - url: http://localhost:8000
    description: Local development orchestrator
paths:
...
  /api/v1/algos/vector/normalize:
    post:
      summary: L2 Vector Unit Normalization (ALGO-VEC-01)
      operationId: normalizeVector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
                eps:
                  type: number
                  default: 1e-12
      responses:
        '200':
          description: L2 normalized unit vector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/center:
    post:
      summary: Corpus Mean Centering & Anisotropy Removal (ALGO-VEC-02)
      operationId: centerVectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vectors]
              properties:
                vectors:
                  type: array
                  items:
                    type: array
                    items: { type: number }
      responses:
        '200':
          description: Centered batch vectors
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/layer-norm:
    post:
      summary: Layer Normalization & Dimension Standardization (ALGO-VEC-03)
      operationId: layerNormVector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
                gamma:
                  type: array
                  items: { type: number }
                beta:
                  type: array
                  items: { type: number }
                eps:
                  type: number
                  default: 1e-5
      responses:
        '200':
          description: Layer normalized vector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/scale:
    post:
      summary: Min-Max and Z-Score Vector Scaling (ALGO-VEC-04)
      operationId: scaleVector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
                min_val:
                  type: number
                  default: 0.0
                max_val:
                  type: number
                  default: 1.0
                method:
                  type: string
                  enum: [minmax, zscore]
                  default: minmax
      responses:
        '200':
          description: Scaled feature vector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/slice:
    post:
      summary: Matryoshka Representation Learning Slicing (ALGO-VEC-05)
      operationId: sliceVector
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
                target_dim:
                  type: integer
                  default: 64
                renormalize:
                  type: boolean
                  default: true
      responses:
        '200':
          description: Sliced and renormalized lower-dimensional vector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/quantize/scalar:
    post:
      summary: Uniform Scalar Quantization SQ8/SQ4 (ALGO-VEC-06)
      operationId: quantizeScalar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
                bits:
                  type: integer
                  enum: [4, 8]
                  default: 8
      responses:
        '200':
          description: Quantized integer values and scaling metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/quantize/binary:
    post:
      summary: 1-Bit Binary Quantization (ALGO-VEC-07)
      operationId: quantizeBinary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [vector]
              properties:
                vector:
                  type: array
                  items: { type: number }
      responses:
        '200':
          description: 1-bit binary representation and packed byte hex
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/pool:
    post:
      summary: Sequence Token Pooling Engine (ALGO-VEC-08)
      operationId: poolTokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token_embeddings]
              properties:
                token_embeddings:
                  type: array
                  items:
                    type: array
                    items: { type: number }
                attention_mask:
                  type: array
                  items: { type: integer }
                pooling_strategy:
                  type: string
                  enum: [mean, cls, last]
                  default: mean
      responses:
        '200':
          description: Pooled sequence representation vector
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
  /api/v1/algos/vector/chunk:
    post:
      summary: Text Chunking and Semantic Breakpoints (ALGO-VEC-09)
      operationId: chunkText
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text:
                  type: string
                max_chunk_size:
                  type: integer
                  default: 200
                overlap:
                  type: integer
                  default: 40
      responses:
        '200':
          description: Segmented text chunks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Envelope_AlgoResponse'
...
components:
  schemas:
    ResponseMeta:
      type: object
      required:
        - requestId
        - correlationId
        - timestamp
        - apiVersion
        - status
        - trace_id
      properties:
        requestId:
          type: string
          example: "req-1700000000000-a1b2c3"
        correlationId:
          type: string
          example: "corr-1700000000000-x9y8z7"
        causationId:
          type: string
          nullable: true
        timestamp:
          type: string
          example: "2026-10-05T14:00:00.000Z"
        executionTimeMs:
          type: integer
          example: 12
        apiVersion:
          type: string
          example: "v1"
        status:
          type: string
          enum: [success, error]
        trace_id:
          type: string
          example: "0af7651916cd43dd8448eb211c80319c"
        auditLogId:
          type: string
          nullable: true
        operationId:
          type: string
          nullable: true
        deprecatedFields:
          type: array
          items: { type: string }
          nullable: true
        supportedVersions:
          type: array
          items: { type: string }
        staleness:
          type: object
          nullable: true
        pagination:
          type: object
          nullable: true
        links:
          type: object
          nullable: true
    ErrorContainer:
      type: object
      required:
        - code
        - message
        - retryable
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        retryable:
          type: boolean
          default: false
        details:
          type: array
          items:
            type: object
    ErrorItem:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "VALIDATION_FAILED"
        message:
          type: string
          example: "One or more payload validation checks failed."
        details:
          type: object
...
    Envelope_AlgoResponse:
      type: object
      required: [success, statusCode, meta, data]
      properties:
        success: { type: boolean, example: true }
        statusCode: { type: integer, example: 200 }
        data: { type: object }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
        error: { $ref: '#/components/schemas/ErrorContainer' }
        errors: { type: array, items: { $ref: '#/components/schemas/ErrorItem' }, nullable: true }
...